fix: isolate Workday native helper entry point
